Refactored authentication service call to its own class.
