+++ src/main/java/com/two/serviceusers/users/UserDao.java
@@ -1,8 +1,5 @@
 package com.two.serviceusers.users;
 
-import com.two.http_api.api.AuthenticationServiceApi;
-import com.two.http_api.model.Tokens;
-import com.two.http_api.model.User;
 import lombok.AllArgsConstructor;
 import org.jooq.DSLContext;
 import org.jooq.generated.tables.records.UserRecord;
@@ -18,7 +15,6 @@ import static org.jooq.generated.Tables.USER;
 public class UserDao {
 
     private final DSLContext ctx;
-    private final AuthenticationServiceApi authenticationServiceApi;
     private static final Logger logger = LoggerFactory.getLogger(UserDao.class);
 
     /**
@@ -39,17 +35,5 @@ public class UserDao {
         logger.info("Successfully stored user in DB with generated UID: {}.", userRecord.getUid());
         return userRecord.getUid();
     }
-
-    /**
-     * @return JWTs from the Authentication Service if the storage was successful.
-     * TODO: Some error handling for WebClient ResponseException
-     */
-    Tokens storeCredentials(int uid, String password) {
-        logger.info("Messaging the Authentication Service to store the users credentials.");
-        return this.authenticationServiceApi.storeCredentialsAndGenerateTokens(
-                new User.Credentials(uid, password)
-        );
-    }
-
 
 }
