+++ src/main/java/com/two/serviceusers/users/UserService.java
@@ -1,6 +1,7 @@
 package com.two.serviceusers.users;
 
 import com.two.http_api.model.Tokens;
+import com.two.serviceusers.authentication.AuthenticationDao;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,6 +14,7 @@ import org.springframework.stereotype.Service;
 public class UserService {
 
     private final UserDao userDao;
+    private final AuthenticationDao authenticationDao;
     private static final Logger logger = LoggerFactory.getLogger(UserService.class);
 
     /**
@@ -24,7 +26,7 @@ public class UserService {
             int uid = this.userDao.storeUser(userRegistration);
 
             logger.info("Storing user credentials.");
-            return this.userDao.storeCredentials(uid, userRegistration.getPassword());
+            return this.authenticationDao.storeCredentials(uid, userRegistration.getPassword());
         } catch (DuplicateKeyException e) {
             logger.warn("The user already exists.", e);
             throw new UserExistsException(userRegistration.getEmail());
